extract common exporter fns
