+++ src/js/exporters/common.js
@@ -1,0 +1,13 @@
+// data functions
+const boardFileImageSize = boardFileData =>
+  (boardFileData.aspectRatio >= 1)
+    ? [900 * boardFileData.aspectRatio, 900]
+    : [900, 900 / boardFileData.aspectRatio]
+
+const msecsToFrames = (fps, value) =>
+  (fps/1000) * value
+
+module.exports = {
+  boardFileImageSize,
+  msecsToFrames
+}
