+++ src/js/exporters/final-cut-pro.js
@@ -1,13 +1,5 @@
 const util = require('../utils')
-
-// data functions
-const boardFileImageSize = boardFileData =>
-  (boardFileData.aspectRatio >= 1)
-    ? [900 * boardFileData.aspectRatio, 900]
-    : [900, 900 / boardFileData.aspectRatio]
-
-const msecsToFrames = (fps, value) =>
-  (fps/1000) * value
+const { boardFileImageSize, msecsToFrames } = require('./common')
 
 // fcp templating
 const clipItem = data =>
